refactor: android 전용 유틸 theme/android/utils로 이전 Related to: #117
