+++ src/main/java/com/komentum/theme/android/dto/AndroidComponentDto.java
@@ -1,7 +1,7 @@
 package com.komentum.theme.android.dto;
 
-import com.komentum.theme.component.domain.ComponentType;
-import com.komentum.theme.component.domain.DesignComponent;
+import com.komentum.designcomponent.domain.ComponentType;
+import com.komentum.designcomponent.domain.DesignComponent;
 import com.komentum.theme.theme.domain.ThemeImage;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
